+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
-- Add unstable `--branch` flag to include branch coverage. ([#356](https://github.com/taiki-e/cargo-llvm-cov/pull/356))
+- Add unstable `--branch` flag to enable branch coverage. ([#356](https://github.com/taiki-e/cargo-llvm-cov/pull/356))
 
 ## [0.6.7] - 2024-03-10