+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [0.6.8] - 2024-03-16
+
 - Add unstable `--branch` flag to enable branch coverage. ([#356](https://github.com/taiki-e/cargo-llvm-cov/pull/356))
 
 ## [0.6.7] - 2024-03-10
@@ -622,7 +624,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/cargo-llvm-cov/compare/v0.6.7...HEAD
+[Unreleased]: https://github.com/taiki-e/cargo-llvm-cov/compare/v0.6.8...HEAD
+[0.6.8]: https://github.com/taiki-e/cargo-llvm-cov/compare/v0.6.7...v0.6.8
 [0.6.7]: https://github.com/taiki-e/cargo-llvm-cov/compare/v0.6.6...v0.6.7
 [0.6.6]: https://github.com/taiki-e/cargo-llvm-cov/compare/v0.6.5...v0.6.6
 [0.6.5]: https://github.com/taiki-e/cargo-llvm-cov/compare/v0.6.4...v0.6.5